% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-compatibility.R
\name{as_}
\alias{as_}
\alias{as_actel}
\alias{as_glatos}
\title{Package conversion routines}
\usage{
as_actel(map, detections, moorings)

as_glatos(detections)
}
\arguments{
\item{map}{SpatRaster}

\item{detections, moorings}{datasets}
}
\description{
This routines convert \code{\link{patter}} \code{\link{data.table}}s into formats used by other packages, such as \code{actel} (\code{RSP}) and \code{glatos}.
}
\details{
\itemize{
\item Convert patter to glatos format
\item Convert \code{detections} \code{\link{data.table}} to \code{actel} format (for RSP)
}
}
