% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-compatibility.R
\name{as_trackyverse}
\alias{as_trackyverse}
\alias{as_actel}
\alias{as_glatos}
\title{Package conversion routines}
\usage{
as_actel(map, detections, moorings)

as_glatos(detections)
}
\arguments{
\item{map}{A \code{\link{SpatRaster}} that defines the study area. This is assumed to define the CRS of the coordinates in \code{moorings}.}

\item{detections, moorings}{\code{\link{patter}} \code{\link{data.table}}s.
\itemize{
\item \code{detections} is a \code{\link{data.table}} of acoustic detections (see \code{\link[patter:datasets-mefs]{patter::dat_detections}}).
\item \code{moorings} is a \code{\link{data.table}} of receiver positions (see \code{\link[patter:datasets-mefs]{patter::dat_moorings}}).
}}
}
\value{
\itemize{
\item \code{\link[=as_actel]{as_actel()}} returns an output from \code{\link[actel:explore]{actel::explore()}};
\item \code{\link[=as_glatos]{as_glatos()}} returns a \code{\link{data.frame}};
}
}
\description{
This routines convert \code{\link{patter}} \code{\link{data.table}}s into formats used by other packages, such as \code{actel} (\code{RSP}) and \code{glatos}.
}
\details{
\itemize{
\item \code{\link[=as_actel]{as_actel()}} converts a \code{detections} \code{\link{data.table}} to \code{actel} \code{explore} format. This is used in \code{\link[=constructor_rsp]{constructor_rsp()}}.
\item \code{\link[=as_glatos]{as_glatos()}} converts a \code{detections} \code{\link{data.table}} to a \code{glatos} \code{det} \code{\link{data.frame}} (e.g., as required in \code{\link[glatos:false_detections]{glatos::false_detections()}}).
}
}
\author{
Edward Lavender
}
