% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cat.R
\name{cat_}
\alias{cat_}
\alias{cat_line}
\alias{cat_iter}
\alias{cat_init}
\title{Utilities: \verb{cat_*()} helpers}
\usage{
cat_line()

cat_iter(t)

cat_init(t)
}
\description{
Utilities: \verb{cat_*()} helpers
}
\keyword{internal}
