% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coffee.R
\name{coffee}
\alias{coffee}
\title{Take a coffee break}
\usage{
coffee(
  computer = Sys.info()["nodename"],
  folder = file.path(tempdir(), "coffee"),
  interval = 2 * 60 * 60,
  duration = 15 * 60
)
}
\description{
Pause and rest the computer
}
\details{
Only one interval & duration is currently supported
}
\author{
Edward Lavender
}
\keyword{internal}
