% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.lapply-interfaces.R
\name{config_particle}
\alias{config_particle}
\title{Configuration options for iterative workflows}
\usage{
config_particle(.cl)
}
\arguments{
\item{.cl}{An \code{integer} that defines the number of cores for a parallelised operation.}
}
\description{
These functions provide additional configure options for selected routines.
}
\details{
\itemize{
\item \code{\link[=config_particle]{config_particle()}} configures sockets for \code{\link[=estimate_coord_particle]{estimate_coord_particle()}}. The function connects to \code{Julia} on each socket (assuming \code{JULIA} options are set globally) and checks for CPU oversubscription.
}
}
\author{
Edward Lavender
}
