% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\name{dat_gebco_tm}
\alias{dat_gebco_tm}
\title{Datasets}
\usage{
dat_gebco_tm()
}
\description{
These functions load selected datasets.
}
\details{
\itemize{
\item \code{\link[=dat_gebco_tm]{dat_gebco_tm()}} loads a \code{TransitionLayer} object from \code{actel::transitionLayer()}. This is predefined to streamline examples.
}
}
\author{
Edward Lavender
}
