% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-map-constructors.R
\name{estimate_map-constructor}
\alias{estimate_map-constructor}
\alias{constructor_map_pou}
\alias{constructor_map_dens}
\alias{constructor_map_dbbmm}
\title{Estimate map: constructors}
\usage{
constructor_map_pou(sim, datasets, verbose, ...)

constructor_map_dens(sim, datasets, verbose, ...)

constructor_map_dbbmm(sim, datasets, verbose, ...)
}
\arguments{
\item{sim, datasets, verbose, ...}{Constructor arguments (see \code{\link{estimate_coord-constructor}}).}
}
\value{
\verb{constructor_*()} functions return a named \code{list} of arguments for an algorithm (e.g., \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}).
}
\description{
\verb{estimate_map_*()} \code{constructor} functions construct a named \code{list} of arguments for an \verb{estimate_map_*()} function.
}
\details{
All \code{constructor} functions have a common signature (see \code{\link{estimate_coord-constructor}}).

For \verb{constructor_map_*()} functions, datasets should include \code{map} and \code{coord} elements. A \code{dataset$coordinate()} element may also be required (see \code{\link[=get_dataset_coord]{get_dataset_coord()}}).

For full customisation, a custom \verb{constructor_*()} function is required.
}
\author{
Edward Lavender
}
