% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-constructors.R
\name{get_}
\alias{get_}
\alias{get_dataset_detections}
\alias{get_dataset_moorings}
\alias{get_dataset_t.layer}
\alias{get_parameter}
\title{Estimate coordinates: get \code{constructor} inputs}
\usage{
get_dataset_detections(sim, datasets)

get_dataset_moorings(sim, datasets)

get_dataset_t.layer(sim, datasets)

get_parameter(sim, parameter)
}
\arguments{
\item{sim, datasets}{Arguments inherited from \code{constructor} in \code{\link[=estimate_coord]{estimate_coord()}}.}
}
\description{
These internal functions get \code{constructor} inputs (i.e., parameters or datasets).
}
\details{
These functions simply extract required inputs from \code{sim} or \code{datasets} with appropriate checks.
}
\author{
Edward Lavender
}
\keyword{internal}
