% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_input}
\alias{get_input}
\alias{get_dataset_map}
\alias{get_dataset_detections}
\alias{get_dataset_moorings}
\alias{get_dataset_t.layer}
\alias{get_parameter}
\title{Estimate coordinates: get \code{algorithm} inputs}
\usage{
get_dataset_map(sim, datasets)

get_dataset_detections(sim, datasets)

get_dataset_moorings(sim, datasets)

get_dataset_t.layer(sim, datasets)

get_parameter(sim, parameter)
}
\arguments{
\item{sim, datasets}{Arguments inherited from \code{constructor} in \code{\link[proj.lapply:workflow]{proj.lapply::workflow()}}.}

\item{parameter}{A \code{character} that defines the name of a column in \code{sim}.}
}
\description{
These functions get \code{algorithm} inputs (i.e., parameters or datasets).
}
\details{
These functions simply extract required inputs from \code{sim} or \code{datasets} with appropriate checks:
\itemize{
\item \verb{get_dataset_*()} functions extract a \code{dataset} \strong{by position} via \code{datasets[[sim$unit_id]]};
\item \code{get_parameter()} functions extract a parameter value via \code{sim[[parameter]]};
}

These functions are used in \code{\link{constructor}} functions.

They are available to users for use in custom \code{\link{constructor}} functions.
}
\author{
Edward Lavender
}
