% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-iter.R
\name{lapply_estimate_coord}
\alias{lapply_estimate_coord}
\alias{lapply_estimate_coord_coa}
\alias{lapply_estimate_coord_rsp}
\alias{lapply_estimate_coord_patter}
\title{Estimate coordinates}
\usage{
lapply_estimate_coord_coa(
  iteration,
  map,
  datasets,
  constructor = constructor_coa,
  log.coffee = NULL,
  log.txt = NULL,
  verbose = TRUE
)

lapply_estimate_coord_rsp(
  iteration,
  map,
  datasets,
  constructor = constructor_rsp,
  log.coffee = NULL,
  log.txt = NULL,
  verbose = TRUE
)

lapply_estimate_coord_patter(
  iteration,
  map,
  datasets,
  constructor,
  log.coffee = NULL,
  log.txt = NULL,
  verbose = TRUE
)
}
\arguments{
\item{iteration}{An iteration \code{\link{data.table}}.
\itemize{
\item Each row corresponds to one algorithm run and is passed to \code{constructor()}.
\item Required columns depend on the \code{constructor()} function but generally include run-specific parameters (such as \code{delta_t} for the \code{\link[patter:coa]{patter::coa()}}) and file paths.
}}

\item{map}{A map that defines the study area, passed to \code{constructor()}.
\itemize{
\item For \code{\link[=estimate_coord_coa]{estimate_coord_coa()}} and \code{\link[=estimate_coord_rsp]{estimate_coord_rsp()}} \code{map} is a \code{\link{SpatRaster}}.
\item For \code{\link[=estimate_coord_patter]{estimate_coord_patter()}}, \code{map} can be a \code{\link{SpatRaster}} or a \code{character} file path.
}}

\item{datasets}{(optional) A named \code{list} of datasets, passed to \code{constructor()}.}

\item{constructor}{A \code{\link{constructor}} \code{function} that constructs a named \code{list} of arguments for the algorithm routine (e.g., \code{\link[patter:coa]{patter::coa()}}).
\itemize{
\item For \code{\link[=algorithm_coa]{algorithm_coa()}} and \code{\link[=algorithm_rsp]{algorithm_rsp()}}, default constructors are provided.
\item For \code{\link[=algorithm_particle]{algorithm_particle()}}, a require a user-defined constructor is required.
}

For function requirements, see \code{\link{constructor}}.}

\item{verbose}{User output control options.
\itemize{
\item \code{verbose} is a \code{logical} variable that defines whether or not to output user messages to the console or \code{log.txt} if specified.
}}

\item{(optional)}{log.coffee,log.txt Log options.
\itemize{
\item \code{log.coffee} is a \code{character} that defines the directory in which to record the timing of 'coffee breaks', during which time the computer is given a rest (see \code{\link[=coffee]{coffee()}}). Use \code{NULL} to suppress coffee breaks.
\item If \code{verbose = TRUE}, \code{log.txt} is a \code{character} that defines a path to a \code{log.txt} file in which output user messages are recorded.
}}
}
\description{
These functions iteratively estimate coordinates using selected algorithms, such as the COA, RSP and particle algorithms (i.e., \code{\link[=coa_algorithm]{coa_algorithm()}}, \code{\link[=rsp_algorithm]{rsp_algorithm()}} and \code{\link[=particle_algorithm]{particle_algorithm()}}).
}
\details{
\verb{lapply_estimate_coord_*()} wraps the internal function \code{\link[=lapply_estimate_coord]{lapply_estimate_coord()}}. This function iterate over rows in \code{iteration} via \code{\link[=cl_lapply]{cl_lapply()}} and applies one of the following internal routines:
\itemize{
\item \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}
\item \code{\link[=estimate_coord_rsp]{estimate_coord_rsp()}}
\item \code{\link[=estimate_coord_patter]{estimate_coord_patter()}}
}

Internally, these routines wrap \code{\link[=estimate_coord]{estimate_coord()}}. This function:
\itemize{
\item Constructs a named \code{list} of arguments for an algorithm function, via a \code{\link{constructor}}, given information in the \code{iteration} row, \code{map} and \code{datasets};
\item Passes the arguments to an algorithm function, such as:
\itemize{
\item \code{\link[=algorithm_coa]{algorithm_coa()}}
\item \code{\link[=algorithm_rsp]{algorithm_rsp()}}
\item \code{\link[=algorithm_particle]{algorithm_particle()}});
}
\item Implements error handling;
\item Records estimated coordinates and call statistics (such as computation time);
\item Optionally writes outputs to disk or returns them in memory
}
}
\author{
Edward Lavender
}
