% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_qplot.R
\name{lapply_qplot}
\alias{lapply_qplot}
\alias{lapply_qplot_coord}
\alias{lapply_qplot_ud}
\title{Quick plots}
\usage{
lapply_qplot_coord(
  .iteration,
  .datasets = list(),
  .n_plot = 4L,
  .n_coord = 10000L,
  .map,
  ...
)

lapply_qplot_ud(.iteration, .datasets = list(), .n_plot = 4L, ...)
}
\arguments{
\item{.iteration}{The iteration \code{\link{data.table}} (see \code{constructor} functions). If \code{.iteration} contains a \code{file_coord} or \code{file_ud} column, map layers (coordinates or utilisation distributions) are read from file; otherwise, they are extracted from \code{.datasets}.}

\item{.datasets}{A named \code{list} of datasets, with one element for each \code{iteration}, required if \code{iteration$file_output} is not specified:
\itemize{
\item For \code{\link[=lapply_qplot_coord]{lapply_qplot_coord()}}, \code{.datasets} should contain \code{coord} and (optional) \code{coordinates} elements (see \code{\link[=get_dataset_coord]{get_dataset_coord()}}). The extracted coordinates should comprise a \code{\link{data.table}} of \code{x} and \code{y} coordinates for this function.
\item For \code{\link[=lapply_qplot_map]{lapply_qplot_map()}}, \code{.datasets} should contain a \code{ud} element  (see \code{\link[=get_dataset_ud]{get_dataset_ud()}}).
}}

\item{.n_plot}{An \code{integer} that defines the number of plots.}

\item{.n_coord}{For \code{\link[=lapply_qplot_coord]{lapply_qplot_coord()}}, \code{.n_coord} is an \code{integer} that defines the maximum number of coordinate pairs to plot.}

\item{.map}{For \code{\link[=lapply_qplot_coord]{lapply_qplot_coord()}}, \code{.map} is a spatial layer, plotted via \code{\link[terra:plot]{terra::plot()}}, onto which coordinates are added.}
}
\description{
These functions quickly plot maps, reading map layers from memory or from file.
}
\details{
This functions produce quick multi-panel maps for with \code{.n_plot} panels.
\itemize{
\item If \code{.iteration$file_output} is specified, map layers are read into memory;
\item Otherwise, map layers are extracted from \code{.datasets};
\item \code{\link[=lapply_qplot_coord]{lapply_qplot_coord()}} plots background maps with coordinates ontop;
\item \code{\link[=lapply_qplot_ud]{lapply_qplot_ud()}} plots utilisation distributions;
\item Map titles are defined by the row index in \code{.iteration};
}
}
\author{
Edward Lavender
}
