% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter.workflows-package.R
\docType{package}
\name{patter.workflows}
\alias{patter.workflows-package}
\alias{patter.workflows}
\title{\code{\link{patter.workflows}}}
\description{
Workflows for the `patter` R package.
}
\examples{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/patter.workflows/}
  \item Report bugs at \url{https://github.com/edwardlavender/patter.workflows/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
