% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-algorithms.R
\name{algorithm}
\alias{algorithm}
\alias{algorithm_coa}
\alias{algorithm_particle}
\title{Estimate coordinates: algorithms}
\usage{
algorithm_coa(
  .map,
  .detections,
  .moorings = NULL,
  .delta_t,
  .split = NULL,
  .plot_weights = TRUE,
  ...,
  .one_page = TRUE
)

algorithm_particle(sim, map, datasets, constructor)

algorithm_particle(sim, map, datasets, constructor)
}
\description{
These functions are algorithms.
}
\details{
\itemize{
\item \code{\link{algorithm_coa}} wraps \code{patter::coa()}
\item \code{\link{algorithm_rsp}} is \code{\link[RSP:runRSP]{RSP::runRSP()}}
\item \code{\link{algorithm_particle}} wraps \code{\link[patter:pf_filter]{patter::pf_filter()}} and \code{\link[patter:pf_smoother_two_filter]{patter::pf_smoother_two_filter()}} (via \code{\link[=pf_filter_wrapper]{pf_filter_wrapper()}} and \code{\link[=pf_smoother_wrapper]{pf_smoother_wrapper()}});
}
}
\author{
Edward Lavender
}
\keyword{internal}
