% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cat.R
\name{cat_}
\alias{cat_}
\alias{cat_time}
\alias{cat_line}
\alias{cat_row}
\alias{cat_next}
\title{Utilities: \verb{cat_*()} helpers}
\usage{
cat_time()

cat_line()

cat_row(index)

cat_next(index, verbose)
}
\description{
Utilities: \verb{cat_*()} helpers
}
\keyword{internal}
