% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.lapply-interfaces.R
\name{config}
\alias{config}
\alias{particle_startup}
\title{Additional routines for iterative workflows with \code{proj.lapply}}
\usage{
particle_startup(.sim, .cl)
}
\arguments{
\item{.sim, .cl}{Arguments inherited from \code{\link[proj.lapply:cl_lapply_workflow]{proj.lapply::cl_lapply_workflow()}}.}
}
\description{
These functions provide additional configure options for selected routines.
}
\details{
\itemize{
\item \code{\link[=particle_startup]{particle_startup()}} configures sockets for \code{\link[=estimate_coord_particle]{estimate_coord_particle()}}. The function connects to \code{Julia} on each socket (assuming \code{JULIA} options are set globally) and checks for CPU oversubscription.
}
}
\author{
Edward Lavender
}
