% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-constructors.R
\name{constructor}
\alias{constructor}
\alias{constructor_coa}
\alias{constructor_rsp}
\title{Estimate coordinates: constructors}
\usage{
constructor_coa(sim, map, datasets)

constructor_rsp(sim, map, datasets)
}
\description{
\code{constructor} functions construct a named \code{list} of arguments for an algorith implementation.
}
\details{
Default \code{constructors} are provided for the COA, RSP and AC particle algorithms.

For \code{\link[=constructor_coa]{constructor_coa()}}:
\itemize{
\item \code{sim} should include a \code{delta_t} column
\item \code{datasets} should include \code{detections_by_unit} and \code{moorings}
}

For \code{\link[=constructor_rsp]{constructor_rsp()}}:
\itemize{
\item \code{sim} should include an \code{er.ad} column
\item \code{datasets} should include \code{detections_by_unit}, \code{moorings} and \code{t.layer}
}

For particle algorithms, constructors have to be defined manually.
}
\author{
Edward Lavender
}
\keyword{internal}
