% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-constructors.R
\name{constructor}
\alias{constructor}
\alias{constructor_coa}
\alias{constructor_rsp}
\title{Estimate coordinates: constructors}
\usage{
constructor_coa(sim, map, datasets, verbose)

constructor_rsp(sim, map, datasets, verbose)
}
\arguments{
\item{sim, map, datasets, verbose}{Arguments inherited from \code{\link[=lapply_estimate_coord]{lapply_estimate_coord()}}:
\itemize{
\item \code{sim} is a \code{\link{data.table}} row;
\item Other arguments are as described for \code{\link[=lapply_estimate_coord]{lapply_estimate_coord()}};
}}
}
\value{
\verb{constructor_*()} functions return a named \code{list} of arguments for an algorithm (e.g., \code{\link[=algorithm_coa]{algorithm_coa()}}).
}
\description{
\code{constructor} functions construct a named \code{list} of arguments for an algorithm implementation.
}
\details{
The following \code{constructors} are built in to \code{\link{patter.workflows}}:
\itemize{
\item \code{\link[=constructor_coa]{constructor_coa()}} assembles a named list of arguments for \code{\link[=algorithm_coa]{algorithm_coa()}};
\item \code{\link[=constructor_rsp]{constructor_rsp()}} assembles a named list of arguments for \code{\link[=algorithm_rsp]{algorithm_rsp()}};
}

For \code{\link[=constructor_coa]{constructor_coa()}}:
\itemize{
\item \code{sim} should include the following columns:
\itemize{
\item \code{unit_id}: an integer that defines the individual/block ID;
\item \code{delta_t}: the time interval over which to calculate centres of activity (see \code{\link[=algorithm_coa]{algorithm_coa()}});
}
\item \code{datasets} should include the following elements:
\itemize{
\item \code{detections_by_unit}: a \code{list} of acoustic detection \code{\link{data.table}}s (see \code{\link[=algorithm_coa]{algorithm_coa()}}) for each \code{unit_id};
\item \code{moorings}: a \code{moorings} data.table (see \code{\link[=algorithm_coa]{algorithm_coa()}});
}
}

For \code{\link[=constructor_rsp]{constructor_rsp()}}:
\itemize{
\item \code{sim} should include the following columns:
\itemize{
\item \code{unit_id}: see above;
\item \code{er.ad}: the \code{er.ad} parameter (see \code{\link[=algorithm_rsp]{algorithm_rsp()}});
}
\item \code{datasets} should include the following elements:
\itemize{
\item \code{detections_by_unit}: see above;
\item \code{moorings}: see above;
\item \code{t.layer}: the transition matrix (see \code{\link[=algorithm_rsp]{algorithm_rsp()}})
}
}

Inbuilt constructors are implemented as follows:
\itemize{
\item The dataset(s) for \code{sim$unit_id} (e.g., for a specific individual/time window) are extracted from \code{datasets} \strong{by position} via \verb{get_dataset_*()} functions (e.g., \code{\link[=get_dataset_detections]{get_dataset_detections()}});
\item Supported parameters (e.g., \code{er.ad}) are extracted from \code{sim} via \code{\link[=get_parameter]{get_parameter()}};
\item (Other parameters are left at default settings);
\item A named \code{list} of arguments, including datasets and parameters, for the relevant function (e.g., \code{\link[=algorithm_coa]{algorithm_coa()}}) is returned;
}

For more control, define a custom \verb{constructor_*()} function.

For \code{\link[=algorithm_particle]{algorithm_particle()}} constructors have to be defined manually, following the same function signature.
}
\author{
Edward Lavender
}
