% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-constructors.R,
%   R/estimate-map-constructors.R
\name{estimate_coord-constructor}
\alias{estimate_coord-constructor}
\alias{constructor_coa}
\alias{constructor_rsp}
\alias{constructor_map_pou}
\alias{constructor_map_dens}
\alias{constructor_map_dbbmm}
\title{Estimate coordinates: constructors}
\usage{
constructor_coa(sim, datasets, verbose)

constructor_rsp(sim, datasets, verbose)

constructor_map_pou(sim, datasets, verbose)

constructor_map_dens(sim, datasets, verbose)

constructor_map_dbbmm(sim, datasets, verbose)
}
\arguments{
\item{sim, datasets, verbose}{Constructor arguments.}
}
\value{
\verb{constructor_*()} functions return a named \code{list} of arguments for an algorithm (e.g., \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}).

\verb{constructor_*()} functions return a named \code{list} of arguments for an algorithm (e.g., \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}).
}
\description{
\verb{estimate_coord_*()} \code{constructor} functions construct a named \code{list} of arguments for an \verb{estimate_coord_*()} function.

\verb{estimate_map_*()} \code{constructor} functions construct a named \code{list} of arguments for an \verb{estimate_map_*()} function.
}
\details{
The following \code{constructors} are built in to \code{\link{patter.workflows}}:
\itemize{
\item \code{\link[=constructor_coa]{constructor_coa()}} assembles a named list of arguments for \code{\link[=estimate_coord_coa]{estimate_coord_coa()}};
\item \code{\link[=constructor_rsp]{constructor_rsp()}} assembles a named list of arguments for \code{\link[=estimate_coord_rsp]{estimate_coord_rsp()}};
}

For \code{\link[=constructor_coa]{constructor_coa()}}:
\itemize{
\item \code{sim} should include the following columns:
\itemize{
\item \code{unit_id}: an integer that defines the individual/block ID;
\item \code{delta_t}: the time interval over which to calculate centres of activity (see \code{\link[=estimate_coord_coa]{estimate_coord_coa()}});
}
\item \code{datasets} should include the following elements:
\itemize{
\item \code{detections_by_unit}: a \code{list} of acoustic detection \code{\link{data.table}}s (see \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}) for each \code{unit_id};
\item \code{moorings}: a \code{moorings} data.table (see \code{\link[=estimate_coord_coa]{estimate_coord_coa()}});
}
}

For \code{\link[=constructor_rsp]{constructor_rsp()}}:
\itemize{
\item \code{sim} should include the following columns:
\itemize{
\item \code{unit_id}: see above;
\item \code{er.ad}: the \code{er.ad} parameter (see \code{\link[=estimate_coord_rsp]{estimate_coord_rsp()}});
}
\item \code{datasets} should include the following elements:
\itemize{
\item \code{detections_by_unit}: see above;
\item \code{moorings}: see above;
\item \code{t.layer}: the transition matrix (see \code{\link[=estimate_coord_rsp]{estimate_coord_rsp()}})
}
}

Inbuilt constructors are implemented as follows:
\itemize{
\item The dataset(s) for \code{sim$unit_id} (e.g., for a specific individual/time window) are extracted from \code{datasets} \strong{by position} via \verb{get_dataset_*()} functions (e.g., \code{\link[=get_dataset_detections]{get_dataset_detections()}});
\item Supported parameters (e.g., \code{er.ad}) are extracted from \code{sim} via \code{\link[=get_parameter]{get_parameter()}};
\item (Other parameters are left at default settings);
\item A named \code{list} of arguments, including datasets and parameters, for the relevant function (e.g., \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}) is returned;
}

For more control, define a custom \verb{constructor_*()} function.

For \code{\link[=estimate_coord_particle]{estimate_coord_particle()}} constructors have to be defined manually, following the same function signature.

All \code{constructor} functions have a common signature (see \code{\link{estimate_coord-constructor}}).

For \code{\link[=constructor_map_pou]{constructor_map_pou()}}:
\itemize{
\item \code{datasets} should include the following elements:
\itemize{
\item \code{map}
\item \code{coord}
}
}

For \code{\link[=constructor_map_dens]{constructor_map_dens()}}:
\itemize{
\item \code{sim} should include:
\itemize{
\item \code{sigma}
}
\item \code{datasets} should include the following elements:
\itemize{
\item \code{map}
\item \code{coord}
}
}

For \code{\link{constructor_map_dbbmm}}:
\itemize{
\item \code{datasets} should include the following elements:
\itemize{
\item \code{map}
\item \code{coord}
\item \code{base.raster}
\item \code{UTM}
}
}

For full customisation, a custom \verb{constructor_*()} function is required.
}
\author{
Edward Lavender
}
