% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord.R
\name{estimate_coord}
\alias{estimate_coord}
\title{Estimate coordinates}
\usage{
estimate_coord(sim, map, datasets, constructor, algorithm, coffee, verbose)
}
\arguments{
\item{sim, map, datasets, constructor, algorithm, coffee}{Arguments directly inherited from \code{\link[=lapply_estimate_coord]{lapply_estimate_coord()}}.}

\item{verbose}{A \code{logical} variable that defines whether or not to send use outputs to the console or a \code{.txt} file.}
}
\description{
This internal function is the workhorse of \code{\link[=lapply_estimate_coord]{lapply_estimate_coord()}}. The function wraps an \code{algorithm} function that estimates coordinates.
}
\details{
For details, see the wrapper function \code{\link[=lapply_estimate_coord]{lapply_estimate_coord()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
