% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-coord-algorithms.R
\name{estimate_coord}
\alias{estimate_coord}
\alias{estimate_coord_coa}
\alias{estimate_coord_rsp}
\alias{estimate_coord_particle}
\title{Estimate coordinates: algorithms}
\usage{
estimate_coord_coa(...)

estimate_coord_rsp(...)

estimate_coord_particle(forward, backward, smooth, verbose)
}
\arguments{
\item{...}{Arguments for algorithms defined in other packages:
\itemize{
\item For \code{\link{estimate_coord_coa}}, \code{...} is passed directly to \code{\link[patter:coa]{patter::coa()}};
\item For \code{\link{estimate_coord_rsp}}, \code{...} is passed directly to \code{\link[RSP:runRSP]{RSP::runRSP()}};
}}

\item{forward, backward, smooth, verbose}{Arguments for \code{\link[=estimate_coord_particle]{estimate_coord_particle()}}:
\itemize{
\item \code{forward} is a named \code{list} of arguments for the forward particle filter (\code{\link[=pf_filter]{pf_filter()}});
\item \code{backward} is a named \code{list} of arguments for the backward particle filter (\code{\link[=pf_filter]{pf_filter()}});
\item \code{smooth} is a named \code{list} of arguments for the two-filter smoother (\code{\link[=pf_smoother_two_filter]{pf_smoother_two_filter()}});
}

To implement smoothing, all arguments must be set. Use \code{smooth = list()} to implement smoothing with default arguments.

Use \code{backward = NULL} and \code{smooth = NULL} to suppress smoothing (and only run the forward particle filter).}
}
\value{
All functions return estimated coordinates in function-specific formats:
\itemize{
\item For \code{\link[=estimate_coord_coa]{estimate_coord_coa()}}, see \code{\link[patter:coa]{patter::coa()}};
\item For \code{\link[=estimate_coord_rsp]{estimate_coord_rsp()}}, see \code{\link[RSP:runRSP]{RSP::runRSP()}};
\item For \code{\link[=estimate_coord_particle]{estimate_coord_particle()}}, a \code{named} list with three elements is returned:
\itemize{
\item \code{forward} holds the outputs from the forward particle filter (\code{\link[=pf_filter]{pf_filter()}});
\item \code{backward} holds the outputs from the backward particle filter (\code{\link[=pf_filter]{pf_filter()}}) or is \code{NULL} if this routine was not implemented;
\item \code{smooth} holds the outputs from the two-filter smoother (\code{\link[=pf_smoother_two_filter]{pf_smoother_two_filter()}}) or is \code{NULL} if this routine was not implemented;
}
}
}
\description{
These functions implement algorithms that estimate coordinates.
}
\author{
Edward Lavender
}
